Backend/requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrdict==2.0.1
babel==2.18.0
bce-python-sdk==0.9.63
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.5
cssselect==1.4.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.25.1
fire==0.7.1
Flask==3.1.3
flask-babel==4.0.0
fonttools==4.62.0
fsspec==2026.2.0
future==1.0.0
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-core==2.5.0
google-cloud-firestore==2.24.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
ImageIO==2.37.3
imagesize==2.0.0
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
lap==0.5.13
lazy-loader==0.5
lmdb==1.7.5
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.34.0
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.7.3
paddlepaddle==2.6.2
paddlex==3.4.2
pandas==2.3.3
pdf2docx==0.5.12
pillow==12.1.1
polars==1.38.1
polars-runtime-32==1.38.1
premailer==3.10.0
prettytable==3.17.0
proto-plus==1.27.1
protobuf==3.20.2
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdfium2==5.6.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.2
RapidFuzz==3.14.3
rarfile==4.2
requests==2.32.5
rich==14.3.3
rsa==4.9.1
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
sympy==1.13.1
termcolor==3.3.0
tifffile==2025.5.10
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
visualdl==2.5.3
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
wsproto==1.3.2
